### 2022-08-11 #4671 @danielskatz
Also, @ritamraha, I see your submission note to JOSS

>A paper presenting the algorithms behind this software (Scarlet) has been published in TACAS'22 (https://link.springer.com/chapter/10.1007/978-3-030-99524-9_14). This software adds some features (works on noisy data) on top of the algorithms in the published paper.

Can you explain the differences in more detail?

@adi3 - please be aware of this and see if we need a scope check
