### 2022-08-14 #4671 @ritamraha
> Also, @ritamraha, I see your submission note to JOSS
>
> > A paper presenting the algorithms behind this software (Scarlet) has been published in TACAS'22 (https://link.springer.com/chapter/10.1007/978-3-030-99524-9_14). This software adds some features (works on noisy data) on top of the algorithms in the published paper.
>
> Can you explain the differences in more detail?
>
> @adi3 - please be aware of this and see if we need a scope check

Hi @adi3, Thanks for agreeing to be the editor. The paper at TACAS'22 explains the algorithms for temporal logic inference from samples and does binary classification, i.e., the resulting temporal logic formula precisely separates all the positive traces from the negative traces. In this software, we have implemented these algorithms and also added the feature of doing the same on noisy data, i.e., now a user can input noisy data as samples and mention a threshold on the loss function. Then, the software will output a temporal logic formula that separates the positive traces from the negatives allowing loss up to that threshold.
