### 2022-08-14 #4671 @ritamraha
> @ritamraha - I'll be the editor for this review. To get things going, could you please suggest 5-6 reviewers from [this](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822) list that you consider appropriate to vet this paper? While selecting, you should confirm that the domain expertise of the reviewer matches the subject matter of your paper and software.
>
> Thank you!

Hi, I am suggesting five possible reviewers' names from the list:
- expectopatronum
- hbaniecki
- feliperodri
- Kraks
- mmpourhashem (although his preferred language is JAVA and our software is in python, his expertise matches the keywords from our paper.)

Thanks!
