+++ paper.bib
@@ -153,4 +153,5 @@
   pages     = {263--280},
   publisher = {Springer},
   year      = {2022}
+  doi       = {10.1007/978-3-030-99524-9_14}
 }
