### 2022-08-14 b7dc0e2
added the missing doi in paper.bib
