+++ paper.bib
@@ -152,6 +152,6 @@
   volume    = {13243},
   pages     = {263--280},
   publisher = {Springer},
-  year      = {2022}
-  doi       = {10.1007/978-3-030-99524-9_14}
+  year      = {2022},
+  doi       = {10.1007/978-3-030-99524-9_14},
 }
