### 2022-09-05 #4671 @adi3
@expectopatronum @hbaniecki @feliperodri @Kraks @mmpourhashem - Would any of you be interested in reviewing this paper on robotics? We'd greatly appreciate your help!
