### 2022-09-05 #4671 @mmpourhashem
Can I have more information about the paper?
> On 1401-6-14, at 1:15 PM, Adi Singh ***@***.***> wrote:
>
>
> @expectopatronum <https://github.com/expectopatronum> @hbaniecki <https://github.com/hbaniecki> @feliperodri <https://github.com/feliperodri> @Kraks <https://github.com/Kraks> @mmpourhashem <https://github.com/mmpourhashem> - Would any of you be interested in reviewing this paper on robotics? We'd greatly appreciate your help!
>
