### 2022-09-05 #4671 @adi3
@mmpourhashem - Sure! [Here](https://github.com/openjournals/joss-papers/blob/joss.04671/joss.04671/10.21105.joss.04671.pdf) is the paper draft. Please feel free to ask the author @ritamraha any questions on paper details.
