### 2022-09-05 #4671 @mmpourhashem
I am fine with reviewing the paper. You can send review invitation to ***@***.*** ***@***.***> that is my academic email ID.

> On 1401-6-14, at 1:41 PM, Adi Singh ***@***.***> wrote:
>
>
> @mmpourhashem <https://github.com/mmpourhashem> - Sure! Here <https://github.com/openjournals/joss-papers/blob/joss.04671/joss.04671/10.21105.joss.04671.pdf> is the paper draft. Please feel free to ask the author @ritamraha <https://github.com/ritamraha> any questions on paper details.
>
