### 2022-10-26 #4671 @ritamraha
> @ritamraha as we continue our search for the second reviewer, can you please list a few more names from the reviewers list who bring in relevant expertise on the subject? I believe you should be on the lookout for folks who specialize in mathematics and algorithmic implementations.

@adi3, I hereby mention some more reviewers' names whose expertise seems to be aligned with the paper. We hope that one of the following reviewers agrees to review our software:

- JDRomano2
- sea-bass
- BasTesterink
- cknoll
- Kraks
- ChristopherHaydenTodd
- ankur-gupta
