### 2022-12-26 #4671 @JDRomano2
I can do it.Thanks,-Joe RomanoSent from my iPhoneOn Dec 26, 2022, at 3:46 PM, Adi Singh ***@***.***> wrote:﻿
Hi @JDRomano2 @sea-bass @BasTesterink @cknoll @Kraks @ChristopherHaydenTodd @ankur-gupta - Can you please acknowledge this message: Would any of you be interested in reviewing this paper for the journal? Thanks!
