### 2023-01-06 #5052 @editorialbot
[REVIEW]: Scarlet: Scalable Anytime Algorithms for Learning Fragments of Linear Temporal Logic
**Submitting author:** @ritamraha (<a href="http://orcid.org/0000-0003-1467-1182">Ritam Raha</a>)
**Repository:** https://github.com/rajarshi008/Scarlet
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
