### 2023-02-07 #5052 @JDRomano2
This paper seems useful and potentially suitable for inclusion in JOSS, but there are some significant issues with the repository and some less significant issues with the paper itself. At a minimum, the following need to be resolved (some are more involved edits than others).

Issues with the code repository (https://github.com/rajarshi008/Scarlet/tree/main):

- Importantly, the repository is not structured appropriately for a Python package. If it is meant to be a `pip` package, like the README suggests, the authors should read the Python Packaging overview (https://packaging.python.org/en/latest/overview/) and likely should also follow the tutorial (https://packaging.python.org/en/latest/tutorials/packaging-projects/). The README should state that all they need to do for installation is run `pip install scarlet-ltl` (no need for the `installation.sh` and cloning the repository, etc.).
- The software does not run on Windows. In fact, the Git repository won't even clone successfully on Windows due to issues with filenames. While not all software needs to be compatible with Windows, there is no reason for this software not to be - none of the dependencies that won't work on Windows. It's just a result of giving an example that relies on running a shell script, and the aforementioned issue with filenames.
- The `LICENSE` file is named incorrectly and therefore Github doesn't detect that there is an OSS license. It needs to be named `LICENSE`, `LICENSE.txt`, `LICENSE.md`, etc.
- There are no automated tests. An automated testing framework such as `pytest` or `nose` should be used to test the code. Currently, you simply mention how the user can 'test' the code by running the software using one or more of the benchmark trace files.
- There are no contribution guidelines.

Issues with the manuscript:
- The summary is not high-level enough for JOSS' audience. What are traces, and why would someone want to generate temporal properties over them? You mention that they can be considered explainable models, but an unfamiliar reader has no idea what they are or why they are more explainable. If word count is a concern, I think you could trim some out of the "Related works" section.
- There is some semantic confusion. The paper states that Scarlet provides an API but there is no API documentation or usage instructions. The repository seems to indicate it instead only a command line utility that processes input and writes output.

I'm going to file these as issues on the repository, but it was helpful for me to write it all out like this in a single comment first.
