### 2023-02-07 rajarshi008/Scarlet#2 @JDRomano2
Does not run on Windows
(review comment for https://github.com/openjournals/joss-reviews/issues/5052)

The software does not run on Windows. In fact, the Git repository won't even clone successfully on Windows due to issues with filenames. While not all software needs to be compatible with Windows, there is no reason for this software not to be - none of the dependencies that won't work on Windows. It's just a result of giving an example that relies on running a shell script, and the aforementioned issue with filenames.

An example of the error I get when trying to clone the repository is as follows:

```
Cloning into 'Scarlet'...
remote: Enumerating objects: 2321, done.
remote: Total 2321 (delta 552), reused 721 (delta 502), pack-reused 1544R
Resolving deltas: 100% (1694/1694), done.
error: invalid path 'All_benchmarks/RQ1-benchmarks/generated_benchmarks/TracesFiles/f:01-nw:025-ml:08-0.trace'
fatal: unable to checkout working tree
warning: Clone succeeded, but checkout failed.
You can inspect what was checked out with 'git status'
and retry with 'git restore --source=HEAD :/'
```

The resulting `Scarlet` directory is then empty.
