### 2023-02-07 rajarshi008/Scarlet#3 @JDRomano2
License file named incorrectly; not detected by GitHub
(review comment for https://github.com/openjournals/joss-reviews/issues/5052)

The `LICENSE` file is named incorrectly and therefore Github doesn't detect that there is an OSS license. It needs to be named `LICENSE`, `LICENSE.txt`, `LICENSE.md`, etc.

You'll know Github has detected it successfully if there is a note near the top of the repository's home page stating that it uses the MIT license.
