### 2023-02-07 rajarshi008/Scarlet#5 @JDRomano2
No contribution guidelines
(review comment for https://github.com/openjournals/joss-reviews/issues/5052)

There are no contribution guidelines provided. These can either go in the README under an appropriate heading, or in a separate `CONTRIBUTING.md` file.
