### 2023-02-07 rajarshi008/Scarlet#6 @JDRomano2
Paper summary is not high-level enough for JOSS general audience
(review comment for https://github.com/openjournals/joss-reviews/issues/5052)

The summary is not high-level enough for JOSS' audience. What are traces, and why would someone want to generate temporal properties over them? You mention that they can be considered explainable models, but an unfamiliar reader has no idea what they are or why they are more explainable. If word count is a concern, I think you could trim some out of the "Related works" section.
