### 2023-02-07 rajarshi008/Scarlet#7 @JDRomano2
Semantic confusion in paper - does Scarlet provide an API for reuse?
(review comment for https://github.com/openjournals/joss-reviews/issues/5052)

There is some semantic confusion. The paper states that Scarlet provides an API but there is no API documentation or usage instructions. The repository seems to indicate it instead only a command line utility that processes input and writes output.

Could you please provide further clarification?
