### 2023-02-27 6f17723
changed the license file name and updated README to include python api
