+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "Scarlet-ltl"
-version = "0.0.1"
+version = "0.0.2"
 authors = [
   { name="Ritam Raha", email="ritam.raha18@gmail.com" },
   { name="Rajarshi Roy", email="rajarshi008@gmail.com" },
