+++ README.md
@@ -24,6 +24,8 @@ An example of a trace is `1,0,1;0,0,0` which consists of two letters each of whi
 1,0,0;1,0,0;0,1,0;1,1,0;1,0,0
 0,0,1;1,0,0;1,1,0;1,1,1;1,0,0
 ```
+For Scarlet to work, you must save your input files with the extension  `.trace`.
+
 
 ## Python API
 
@@ -35,38 +37,35 @@ Now, you can install the tool, as python package using pip command as follows:
 python3 -m pip install Scarlet-ltl
 ```
 
-### Create input file
-
-To run Scarlet, you have to create an input file with `.trace` extension in the same directory where `venv` folder is located. The input file format is described in the above section.
-
 ### Run Scarlet on a particular input file
 
 ```
 from Scarlet.ltllearner import LTLlearner
-learner = LTLlearner(input_file = "input_file_name.trace")
+learner = LTLlearner(input_file = input_file_name)
 learner.learn()
 ```
-This will run Scarlet on the input trace file.
+This will run Scarlet on the input_trace_file. Note that the python API does not provide any example input files.
+Hence, you must provide your trace file in a format as described above. 
 
 ### Parameters
 You can call the `LTLlearner` class with additional parameters as follows:
 
-* input_file = the path of the file containing LTL formuas, i.e., `= 'input_file_name.trace'`
+* input_file = the path of the file containing LTL formuas, e.g., `= 'example.trace'`
 * timeout = For specifying the timeout, `default = 900`
-* csvname = the name of the output csv file, i.e.,  `= 'output_file_name.csv'`
+* csvname = the name of the output csv file, e.g.,  `= 'result.csv'`
 * thres = the bound on loss function for noisy data, `default = 0` for perfect classification, has to be a number between zero and one
 
 
-## How to generate trace files from LTL formulas
+### How to generate trace files from LTL formulas
 
 You can also generate trace files from given LTL formulas following the instructions below:
 
-### Install dependencies
+- Install dependencies
 
 For generating benchmarks from a given set of LTL formula, we rely on a python package LTLf2DFA that uses [MONA](https://www.brics.dk/mona/) in its backend. 
 As a result, one needs to install MONA first in order to be able to use this procedure (instructions can be found in the MONA website).
 
-### Create input formula file
+- Create input formula file
 
 For generating benchmarks, you have to create an input file named `formulas.txt` in the same directory where `venv` folder is located. The formula file should contain a list of formulas (in prefix notation) along with the alphabet.
 An example of this file is as follows:
@@ -77,7 +76,7 @@ G(!(p));p
 G(->(q, G(!(p))));p,q
 ```
 
-### Generate trace files from `formulas.txt`
+#### Generate trace files from `formulas.txt`
 
 ```
 from Scarlet.genBenchmarks import SampleGenerator
@@ -85,7 +84,7 @@ generator = SampleGenerator(formula_file= "formulas.txt")
 generator.generate()
 ```
 
-### Parameters
+#### Parameters
 You can call the `SampleGenerator` class with additional parameters as follows:
 
 * formula_file = the path of the file containing LTL formuas, `example = 'formulas.txt'`
