+++ README.md
@@ -24,12 +24,13 @@ An example of a trace is `1,0,1;0,0,0` which consists of two letters each of whi
 1,0,0;1,0,0;0,1,0;1,1,0;1,0,0
 0,0,1;1,0,0;1,1,0;1,1,1;1,0,0
 ```
-For Scarlet to work, you must save your input files with the extension  `.trace`.
+For `Scarlet` to work, you must save your input files with the extension  `.trace`.
 
 
+There are two ways of using `Scarlet` depending on what one prefers. One can use `Scarlet` using its python API  or using its command-line features. Here, we provide the instructions to use both.
 ## Python API
 
-### Installing the tool
+### Installation
 
 Now, you can install the tool, as python package using pip command as follows:
 
@@ -37,15 +38,15 @@ Now, you can install the tool, as python package using pip command as follows:
 python3 -m pip install Scarlet-ltl
 ```
 
-### Run Scarlet on a particular input file
-
+### Basic API Usage
+We now provide the basic usage of the Scarlet python API.
 ```
 from Scarlet.ltllearner import LTLlearner
 learner = LTLlearner(input_file = input_file_name)
 learner.learn()
 ```
-This will run Scarlet on the input_trace_file. Note that the python API does not provide any example input files.
-Hence, you must provide your trace file in a format as described above. 
+Note that the python API does not provide any example input files.
+Hence, you must create your trace file in a format as described above and provide it to the API.
 
 ### Parameters
 You can call the `LTLlearner` class with additional parameters as follows:
@@ -56,27 +57,27 @@ You can call the `LTLlearner` class with additional parameters as follows:
 * thres = the bound on loss function for noisy data, `default = 0` for perfect classification, has to be a number between zero and one
 
 
-### How to generate trace files from LTL formulas
+### Generation of (random) Trace Files from LTL formulas
 
 You can also generate trace files from given LTL formulas following the instructions below:
 
-- Install dependencies
+#### Installation of Dependencies
 
 For generating benchmarks from a given set of LTL formula, we rely on a python package LTLf2DFA that uses [MONA](https://www.brics.dk/mona/) in its backend. 
 As a result, one needs to install MONA first in order to be able to use this procedure (instructions can be found in the MONA website).
 
-- Create input formula file
+#### Creating Input Formula File
 
 For generating benchmarks, you have to create an input file named `formulas.txt` in the same directory where `venv` folder is located. The formula file should contain a list of formulas (in prefix notation) along with the alphabet.
 An example of this file is as follows:
 
 ```
 G(!(p));p
-->(F(q), U(!(p),q));p,q
+|(G(!(p)),F(&(p, F(q))));p,q
 G(->(q, G(!(p))));p,q
 ```
 
-#### Generate trace files from `formulas.txt`
+#### Generating Trace Files for a given Formula File
 
 ```
 from Scarlet.genBenchmarks import SampleGenerator
@@ -92,6 +93,8 @@ You can call the `SampleGenerator` class with additional parameters as follows:
 * trace_lengths = For specifying the length range for each trace in the samples, `default = [(6,6)]`
 * output_folder = For specifying the name of the folder in which samples are generated
 
+The PyPi package for this API can be found in [this link](https://pypi.org/project/Scarlet-ltl/0.0.2/).
+
 
 ## Command Line Usage
 
@@ -102,12 +105,9 @@ git clone https://github.com/rajarshi008/Scarlet.git
 cd Scarlet
 source ./installation.sh
 ```
-You can also install `Scarlet`  as a python package using pip command. To do so, follow [this link](https://pypi.org/project/Scarlet-ltl/0.0.1/).
 
+### Running
 
-## How to run:
-
-### For running the LTL learner
 
[…]
