+++ README.md
@@ -4,8 +4,6 @@
 
 ---
 
-## SCARLET 
-
 We solve the problem of learning LTL formulas from a sample consisting of traces partitioned into positive and negative.
 
 A [paper](https://link.springer.com/chapter/10.1007/978-3-030-99524-9_14) presenting the algorithms behind `Scarlet` was published in TACAS'2022.
