+++ paper.md
@@ -39,16 +39,41 @@ bibliography: paper.bib
 
 # Summary
 
-Linear Temporal Logic (LTL) is a prominent logic for specifying temporal properties over infinite traces, and recently introduced over finite traces. It has become a de facto standard in many fields such as model checking, program analysis, and motion planning for robotics. Over the past five to ten years learning temporal logics (of which LTL is the core) has become an active research area and identified as an important goal in artificial intelligence: it formalises the difficult task of building explainable models from data. Indeed LTL formulas are typically easy to interpret by human users and therefore useful as explanations. The variable free syntax of LTL and its natural inductive semantics make LTL a natural target for building classifiers separating positive from negative traces.
+In the past few years, incorporating data-driven AI techniques in system design has become mainstream, almost in all branches of science and technology.
+Typically, systems powered by AI tend to be rather complex, far beyond human understanding.
+Naturally, one cannot always develop trust in such complex, so-called black-box systems, restricting their widespread use in safety-critical domains.
+
+To generate trust in systems, a standard approach in Explainable AI is to build simple explanations using human-understandable models.
+A number of recent works [@NeiderGavran18] [@Camacho_McIlraith_2019] [@RoyFismanNeider20] have identified models in temporal logic to be both formal and explainable.
+Among temporal logics---Linear Temporal Logic (LTL)---arguably the most widely used temporal logic, has received particular focus due to its ease-of-use.
+LTL is in fact a de facto standard in several fields of computer science, including model checking, program analysis, and motion planning for robotics.
+
+`Scarlet` is one of the most competitive tools for learning explainable models in LTL [@RahaRFN22].
+To learn the explainable models, it relies on positive (or desirable) and negative (or undesirable) executions of the system under consideration.
+Based on the executions, it learns a concise model in LTL that is consistent with the given executions.
+
+Let us consider a concrete example to understand `Scarlet`'s functioning.
+Consider a robot that has been designed to collect wastebin contents in an office-like environment.
+For the sake of the example, in this environment, let there be an office o with a wastebin, a hallway h, and a container c to accumulate the waste.
+The following could be the possible executions of the robot:
+
+h.h.h.h.o.h.c.h
+h.h.h.h.h.c.h.o.h.h
+
+Let the first execution be positive since the robot first collects waste from the office and then accumulates in the container.
+Also, let the second execution be negative since the robot tries to accumulate the waste in the container even before it collects from the office.
+From these executions, `Scarlet` learns a model `F(o and FX c)`, where the `F`-operator stands for "finally" and `X`-operator stands for "next".
+The model, in simple words, expresses that: eventually, the robot should visit the office and then, at a later point visit the container. 
+
 
 # Statement of need
 
-The fundamental problem solved by `Scarlet` is to build an explainable model in the form of an LTL formula from a set of positive and negative traces.
+The fundamental problem solved by `Scarlet` is to build an explainable model in the form of an LTL formula from a set of positive and negative execution traces.
 
-`Scarlet` is a Python package offering an API for tasks related to LTL learning: 
+`Scarlet` is a tool built entirely in Python 3. It can be run using its command-line features or its Python API hosted in PyPi. Its main capabilities include:
 
-* constructing an LTL formula from example traces
-* generating traces from an LTL formula using automata-based techniques (LTLf2DFA and MONA) and random sampling
+* construction of an LTL formula from execution traces,
[…]
