### 2023-04-17 rajarshi008/Scarlet#6 @ritamraha
> (review comment for [openjournals/joss-reviews#5052](https://github.com/openjournals/joss-reviews/issues/5052))
>
> The summary is not high-level enough for JOSS' audience. What are traces, and why would someone want to generate temporal properties over them? You mention that they can be considered explainable models, but an unfamiliar reader has no idea what they are or why they are more explainable. If word count is a concern, I think you could trim some out of the "Related works" section.

Thank you for raising this issue. We have updated the summary to make it more high-level. Currently, it intuitively explains why temporal properties are considered as explainable models via an illustrative example.
