+++ README.md
@@ -2,11 +2,12 @@
   <img src="https://rajarshi008.github.io/assets/images/scarlet-logo.png" width="60%">
  </h1>
 
-We solve the problem of learning LTL formulas from a sample consisting of traces partitioned into positive and negative.
+
+We consider the problem of learning explainable models in Linear Temporal Logic (LTL) using system executions partitioned into positive (or desirable) and negative (or undesirable).
 
 Our paper "[Scalable Anytime Algorithms for Learning Fragments of Linear Temporal Logic](https://link.springer.com/chapter/10.1007/978-3-030-99524-9_14)" presenting the algorithms behind *Scarlet* was published in TACAS'2022.
 
-*Table of Contents*
+**Table of Contents**
 <!-- toc -->
 
 - [Input File format](#input-file-format)
