### 2023-04-17 rajarshi008/Scarlet#5 @ritamraha
> (review comment for [openjournals/joss-reviews#5052](https://github.com/openjournals/joss-reviews/issues/5052))
>
> There are no contribution guidelines provided. These can either go in the README under an appropriate heading, or in a separate `CONTRIBUTING.md` file.

We have added the contribution guidelines in a separate section in README.
