### 2023-04-17 rajarshi008/Scarlet#3 @ritamraha
> (review comment for [openjournals/joss-reviews#5052](https://github.com/openjournals/joss-reviews/issues/5052))
>
> The `LICENSE` file is named incorrectly and therefore Github doesn't detect that there is an OSS license. It needs to be named `LICENSE`, `LICENSE.txt`, `LICENSE.md`, etc.
>
> You'll know Github has detected it successfully if there is a note near the top of the repository's home page stating that it uses the MIT license.

We have fixed the LICENSE file and now the MIT License is recognised by Github.
