### 2023-04-17 rajarshi008/Scarlet#7 @ritamraha
> (review comment for [openjournals/joss-reviews#5052](https://github.com/openjournals/joss-reviews/issues/5052))
>
> There is some semantic confusion. The paper states that Scarlet provides an API but there is no API documentation or usage instructions. The repository seems to indicate it instead only a command line utility that processes input and writes output.
>
> Could you please provide further clarification?

To resolve the semantic confusion, we have clearly mentioned the documentation of our Python API in the README. For bash users, we additionally provide a command line utility. We have mentioned both these options in the paper (summary) and as well as in the README.
