### 2023-04-17 rajarshi008/Scarlet#1 @ritamraha
> (review comment for [openjournals/joss-reviews#5052](https://github.com/openjournals/joss-reviews/issues/5052))
>
> The repository is not structured appropriately for a Python package. If it is meant to be a `pip` package, like the README suggests, the authors should read the Python Packaging overview (https://packaging.python.org/en/latest/overview/) and likely should also follow the tutorial (https://packaging.python.org/en/latest/tutorials/packaging-projects/). The README should state that all they need to do for installation is run `pip install scarlet-ltl` (no need for the `installation.sh` and cloning the repository, etc.).
>
> It looks like maybe this was partially addressed in a separate branch, but it needs to be merged in to `main` (and I think the journal checks will need to be run again, perhaps).

We have fixed this issue by clearly mentioning documentation about installing and using our Python API in README, which is hosted in PyPI.
