### 2023-04-17 #5052 @ritamraha
> @ritamraha can you give us an update on your progress? Please keep your paper reviewers in the loop so they can make way through their checklists. If you fix any issues raised, put that in a message here and tag the corresponding reviewer. Thanks!
>
> @mmpourhashem have you had a chance to review this paper? Please being by generating your checklist and then going through each of the points. For more details, see my [message](https://github.com/openjournals/joss-reviews/issues/5052#issuecomment-1373802810) above. Thank you!

@adi3 I am mentioning my progress here according to @JDRomano2 's reviews. I thank the reviewer for his helpful feedback. I have commented about the fixes to each issue in the tool repository. I am giving an overview of all the fixes here:

1. We have clearly mentioned the usage and installation of our Python API in the README.
2. We have fixed the file names, so it should work on Windows.
3. We have fixed the LICENSE file name, which Github now recognises.
4. In summary, we have explained our contribution in a high-level manner for the general audience.
5. We have added Contribution Guidelines in README.

We are currently working on the last issue - adding automated tests. @JDRomano2, please let me know if there is more feedback based on our changes.
