### 2023-04-28 rajarshi008/Scarlet#2 @JDRomano2
Thank you- I've confirmed that this indeed works correctly now!
(ref: openjournals/joss-reviews#5052)
