+++ tests/test_genbenchmarks.py
@@ -1,0 +1,86 @@
+import pytest
+from Scarlet.genBenchmarks import SampleGenerator
+from Scarlet.sample import Sample, Trace
+import glob
+
+
+formulas = [Formula.convertTextToFormula('G(!(p))'),\
+            Formula.convertTextToFormula('G(->(q, G(!(p))))'),\
+            Formula.convertTextToFormula('F(p)'),\
+            Formula.convertTextToFormula('G(p)'),\
+            Formula.convertTextToFormula('G(->(q, G(p)))')]
+
+@pytest.mark.parametrize("test_num", params)
+def test_genbenchmarks():
+    '''
+    Testing LTLlearner by invoking the verifier to check consistency with the sample
+    '''
+    print(filename, thres)
+    learner = LTLlearner(input_file=filename, timeout=100, thres=thres, csvname=filename.strip('.trace')+'.csv')
+
+    formula = learner.learn()[0]
+
+    sample = Sample(positive=[], negative=[], alphabet=[])
+    sample.readFromFile(filename)
+
+    misclass = 0
+    for w in sample.positive:
+        if w.evaluateFormula(formula,sample.letter2pos) == False:
+            misclass += 1
+
+    for w in sample.negative:
+        if w.evaluateFormula(formula,sample.letter2pos) == True:
+            misclass += 1
+
+
+    misclass_percent = misclass/(len(sample.positive)+len(sample.negative))  
+
+    assert misclass_percent <= thres
+
+
+
+    # Test the find() method with invalid upper_bound
+
+
+#ver true in the test suite
+
+#incr length
+
+#incr width
+
+
+'''
+    bsc.cover_size == {Formula.convertTextToFormula('p'): 8, Formula.convertTextToFormula('F(p)'): 7, Formula.convertTextToFormula('q'): 5, \
+                                    Formula.convertTextToFormula('F(q)'): 8, Formula.convertTextToFormula('!(q)'): 5, \
+                                    Formula.convertTextToFormula('F(!(q))'): 5, Formula.convertTextToFormula('G(!(q))'): 2, \
+                                    Formula.convertTextToFormula('G(p)'): 5, Formula.convertTextToFormula('G(q)'): 5, \
+                                    Formula.convertTextToFormula('G(!(p))'): 3, Formula.convertTextToFormula('X(p)'): 6, \
+                                    Formula.convertTextToFormula('F(X(p))'): 6, Formula.convertTextToFormula('X(!(q))'): 5, \
+                                    Formula.convertTextToFormula('F(X(!(q)))'): 5, Formula.convertTextToFormula('X(q)'): 5, \
+                                    Formula.convertTextToFormula('F(X(q))'): 7, Formula.convertTextToFormula('X(G(!(q)))'): 3, \
+                                    Formula.convertTextToFormula('X(G(p))'): 5, Formula.convertTextToFormula('X(G(q))'): 5, \
+                                    Formula.convertTextToFormula('X(X(!(p)))'): 5, Formula.convertTextToFormula('F(X(X(!(p))))'): 5, \
+                                    Formula.convertTextToFormula('X(X(!(q)))'): 3, Formula.convertTextToFormula('F(X(X(!(q))))'): 3, \
+                                    Formula.convertTextToFormula('X(X(p))'): 5, Formula.convertTextToFormula('F(X(X(p)))'): 5, \
+                                    Formula.convertTextToFormula('X(X(q))'): 7, Formula.convertTextToFormula('F(X(X(q)))'): 7, \
[…]
