+++ tests/test_booleansubsetcover.py
@@ -75,48 +75,3 @@ def test_boolean_set_cover(test_num):
     else:
         assert final_upperbound == upper_bound
 
-    # Test the find() method with invalid upper_bound
-
-
-#ver true in the test suite
-
-#incr length
-
-#incr width
-
-
-'''
-    bsc.cover_size == {Formula.convertTextToFormula('p'): 8, Formula.convertTextToFormula('F(p)'): 7, Formula.convertTextToFormula('q'): 5, \
-                                    Formula.convertTextToFormula('F(q)'): 8, Formula.convertTextToFormula('!(q)'): 5, \
-                                    Formula.convertTextToFormula('F(!(q))'): 5, Formula.convertTextToFormula('G(!(q))'): 2, \
-                                    Formula.convertTextToFormula('G(p)'): 5, Formula.convertTextToFormula('G(q)'): 5, \
-                                    Formula.convertTextToFormula('G(!(p))'): 3, Formula.convertTextToFormula('X(p)'): 6, \
-                                    Formula.convertTextToFormula('F(X(p))'): 6, Formula.convertTextToFormula('X(!(q))'): 5, \
-                                    Formula.convertTextToFormula('F(X(!(q)))'): 5, Formula.convertTextToFormula('X(q)'): 5, \
-                                    Formula.convertTextToFormula('F(X(q))'): 7, Formula.convertTextToFormula('X(G(!(q)))'): 3, \
-                                    Formula.convertTextToFormula('X(G(p))'): 5, Formula.convertTextToFormula('X(G(q))'): 5, \
-                                    Formula.convertTextToFormula('X(X(!(p)))'): 5, Formula.convertTextToFormula('F(X(X(!(p))))'): 5, \
-                                    Formula.convertTextToFormula('X(X(!(q)))'): 3, Formula.convertTextToFormula('F(X(X(!(q))))'): 3, \
-                                    Formula.convertTextToFormula('X(X(p))'): 5, Formula.convertTextToFormula('F(X(X(p)))'): 5, \
-                                    Formula.convertTextToFormula('X(X(q))'): 7, Formula.convertTextToFormula('F(X(X(q)))'): 7, \
-                                    Formula.convertTextToFormula('X(G(!(p)))'): 4, Formula.convertTextToFormula('X(X(G(!(p))))'): 5, \
-                                    Formula.convertTextToFormula('X(X(G(!(q))))'): 3, Formula.convertTextToFormula('X(X(G(p)))'): 5, \
-                                    Formula.convertTextToFormula('X(X(G(q)))'): 7}
-    
-    assert bsc.heap == [( -4.0, Formula.convertTextToFormula('p')), ( -3.31, Formula.convertTextToFormula('F(q)')),\
-                        ( -2.56, Formula.convertTextToFormula('X(X(q))')), ( -2.9, Formula.convertTextToFormula('F(p)')),\
-                        ( -2.49, Formula.convertTextToFormula('X(p)')), ( -2.5, Formula.convertTextToFormula('q')), ( -2.07, Formula.convertTextToFormula('X(q)')),\
-                        ( -2.56, Formula.convertTextToFormula('F(X(q))')), ( -2.07, Formula.convertTextToFormula('G(q)')),\
-                        ( -1.67, Formula.convertTextToFormula('X(X(!p))')), ( -2.07, Formula.convertTextToFormula('!q')),\
-                        ( -1.83, Formula.convertTextToFormula('X(X(p))')), ( -2.33, Formula.convertTextToFormula('F(X(X(q)))')),\
[…]
