+++ tests/test_genbenchmarks.py
@@ -2,85 +2,37 @@ import pytest
 from Scarlet.genBenchmarks import SampleGenerator
 from Scarlet.sample import Sample, Trace
 import glob
+import random
+import os
 
+num_tests = 10
 
-formulas = [Formula.convertTextToFormula('G(!(p))'),\
-            Formula.convertTextToFormula('G(->(q, G(!(p))))'),\
-            Formula.convertTextToFormula('F(p)'),\
-            Formula.convertTextToFormula('G(p)'),\
-            Formula.convertTextToFormula('G(->(q, G(p)))')]
-
-@pytest.mark.parametrize("test_num", params)
-def test_genbenchmarks():
+@pytest.mark.parametrize("test_num", [i for i in range(num_tests)])
+def test_genbenchmarks(test_num):
     '''
-    Testing LTLlearner by invoking the verifier to check consistency with the sample
+    Testing genbenchmarks by checking consistency with the formula
     '''
-    print(filename, thres)
-    learner = LTLlearner(input_file=filename, timeout=100, thres=thres, csvname=filename.strip('.trace')+'.csv')
+    sg = SampleGenerator(formula_file = 'Scarlet/tests/test_benchmarks/formulas.txt',
+                        trace_type = 'trace',
+                        sample_sizes = [(10,10),(50,50)],
+                        trace_lengths = [(6,6)],
+                        output_folder = 'Scarlet/tests/test_generations',
+                        total_num = 1,
+                        gen_method = 'dfa_method'
+                        )
 
-    formula = learner.learn()[0]
+    num_sample = random.randint(10,20)
+    trace_len = random.randint(5,10)
+    sg.sample_sizes = [(num_sample,num_sample)]
+    sg.trace_lengths = [(trace_len,trace_len)]
 
-    sample = Sample(positive=[], negative=[], alphabet=[])
-    sample.readFromFile(filename)
+    sg.generate()
 
-    misclass = 0
-    for w in sample.positive:
-        if w.evaluateFormula(formula,sample.letter2pos) == False:
-            misclass += 1
+    folder_path = 'Scarlet/tests/test_generations/TracesFiles'
+    trace_files = [file for file in os.listdir(folder_path) if file.endswith(".trace")]
 
-    for w in sample.negative:
-        if w.evaluateFormula(formula,sample.letter2pos) == True:
-            misclass += 1
-
-
-    misclass_percent = misclass/(len(sample.positive)+len(sample.negative))  
-
-    assert misclass_percent <= thres
-
-
-
-    # Test the find() method with invalid upper_bound
-
-
-#ver true in the test suite
-
-#incr length
-
-#incr width
-
-
-'''
-    bsc.cover_size == {Formula.convertTextToFormula('p'): 8, Formula.convertTextToFormula('F(p)'): 7, Formula.convertTextToFormula('q'): 5, \
-                                    Formula.convertTextToFormula('F(q)'): 8, Formula.convertTextToFormula('!(q)'): 5, \
-                                    Formula.convertTextToFormula('F(!(q))'): 5, Formula.convertTextToFormula('G(!(q))'): 2, \
-                                    Formula.convertTextToFormula('G(p)'): 5, Formula.convertTextToFormula('G(q)'): 5, \
-                                    Formula.convertTextToFormula('G(!(p))'): 3, Formula.convertTextToFormula('X(p)'): 6, \
-                                    Formula.convertTextToFormula('F(X(p))'): 6, Formula.convertTextToFormula('X(!(q))'): 5, \
[…]

[…]
