+++ requirements.txt
@@ -1,3 +1,4 @@
 lark
 graphviz
 ltlf2dfa
+pytest

[…]
