+++ README.md
@@ -169,7 +169,7 @@ The formula file should contain a list of formulas (in prefix notation) along wi
 ### Testing
 For testing all features of the tool use: `python -m Scarlet/tests` (Installation of *MONA* is required).
 For testing only the features of LTL learning use: `python -m Scarlet/tests/test_learner`.
-For testing only the features of sample generation use: `python -m Scarlet/tests/test_generator` (Installation of *MONA* is required).
+For testing only the features of trace file generation use: `python -m Scarlet/tests/test_generator` (Installation of *MONA* is required).
 
 
 ## Report an issue
