### 2023-07-21 #5052 @Smattr
Thanks for considering me and apologies for the delay in reply. From a quick glance, this certainly looks like something I’m qualified to review. Unfortunately I don’t have the bandwidth for it right now though. If you haven’t found anyone else in ~1.5 months I should have more availability to review then.

Some unofficial first impression comments on the draft paper itself:
* The Summary doesn’t mention Scarlet at all and reads more like background information. Is this not meant to be something like an abstract of the work?
* The mention of artificial intelligence seems a little odd when nothing here seems specific to or especially suited to AI. I guess AI is a hot topic right now, so is this some keyword gaming? (to which I don’t object, but just trying to understand if I’ve missed a connection here.)
* Given the generalised audience of JOSS, this could probably provide more context on the domain (formal methods, proofs, etc). To formal methods practitioners, the utility is going to be clear. But to everyone else, I imagine they’re wondering what you would use things like LTL and SAT for.
