### 2023-09-04 rajarshi008/Scarlet#8 @Smattr
test suite failures
Cloning the latest revision (552bf43c4ff267c4b353bcb6c0fcab5cc046e923) and attempting to run the test suite generates lots of failures:

```bash
$ git clone https://github.com/rajarshi008/Scarlet
$ python3 -m pytest Scarlet/
================================================================================================================================== test session starts ==================================================================================================================================
rootdir: /tmp/tmp.XkaoMMhGjs/foo
collected 178 items

Scarlet/tests/test_generator/test_genbenchmarks.py ..........                                                                                                                                                                                                                     [  5%]
Scarlet/tests/test_learner/test_booleansubsetcover.py ..........                                                                                                                                                                                                                  [ 11%]
Scarlet/tests/test_learner/test_dltl.py FFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF                                                                                                              [ 80%]
Scarlet/tests/test_learner/test_inferLTL.py FFFFFFFFFFFFFFFFFFFF                                                                                                                                                                                                                  [ 92%]
Scarlet/tests/test_learner/test_ltllearner.py ..............                                                                                                                                                                                                                           [100%]

======================================================================================================================================= FAILURES ========================================================================================================================================

…
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[0] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[1] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[2] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[3] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[4] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[5] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[6] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
FAILED Scarlet/tests/test_learner/test_inferLTL.py::test_dltl2Formula[7] - TypeError: dltl2Formula() takes 2 positional arguments but 3 were given
…
```

Maybe this is not the way to run the tests? I have MONA installed, so I don’t think that’s the problem.

(https://github.com/openjournals/joss-reviews/issues/5052)
