### 2023-09-04 rajarshi008/Scarlet#10 @Smattr
Pylint issues
Pylint identifies a number of issues with the Scarlet sources. In particular, with `--errors-only` it finds things like:

```
sample.py:321:18: E0602: Undefined variable 'length_word' (undefined-variable)
```

(https://github.com/openjournals/joss-reviews/issues/5052)
