### 2023-09-04 6a03eb1
remove unnecessary list materialisation
Functions like `tuple` will accept generator expressions, so there is no need to
wrap them in `[]`.
