### 2023-10-02 rajarshi008/Scarlet#10 @ritamraha
We have resolved the issues pointed out by Pylint.
