### 2023-10-02 973b77c
Not equal to None also changed
