+++ Scarlet/decisionTree.py
@@ -1,341 +1,0 @@
-import logging
-import os
-from graphviz import Source
-from Scarlet.formulaTree import *
-
-
-#The decision tree algorithm for combinations of dLTLS
-class DecisionTree:
-
-	def __init__(self, label, thres, leftChild=None, rightChild=None):
-
-		assert(label != None)
-		assert(type(label)==Formula)
-		self.label = label
-		self.left = leftChild
-		self.right = rightChild
-		self.thres = thres
-		self.size = None
-
-	def DTSize(self):
-		'''
-			calculates decision tree size
-		''' 
-		
-		leftSize = 0
-		rightSize = 0
-
-		if self.left != None:
-			if self.left.label != formula_true and self.left.label != formula_false:
-				leftSize = 1 + self.left.DTSize()
-		if self.right != None:
-			if self.right.label != formula_true and self.right.label != formula_false:
-				rightSize = 1 + self.right.DTSize()
-
-		self.size = self.label.treeSize() + leftSize + rightSize  
-		return self.size
-
-	def generate_dot(self):
-		'''
-			generates dot files for the tree  
-		'''
-		tree_queue = [(self,1)]
-		tree_id = {1: self}
-		edges = []
-
-		while tree_queue != []:
-
-			curr_tree, curr_id = tree_queue.pop()
-
-			if curr_tree.left != None:
-				left_id = 2*curr_id 
-				tree_queue.insert(0,(curr_tree.left,left_id))
-				tree_id[left_id] = curr_tree.left
-				edges.append((curr_id,left_id,0))
-			
-			if curr_tree.right != None:
-				right_id = 2*curr_id+1
-				tree_queue.insert(0,(curr_tree.right, right_id))
-				tree_id[right_id] = curr_tree.right
-				edges.append((curr_id,right_id,1))
-
-		dot_str =  "digraph g {\n"
-
-		for i in tree_id:
-			dot_str += ('{} [label="{}"]\n'.format(i, tree_id[i].label))
-		for edge in edges:
-			if edge[2]:
-				dot_str += ('{} -> {}\n [style= {}]'.format(edge[0],edge[1],'dashed'))
-			else:
-				dot_str += ('{} -> {}\n [style= {}]'.format(edge[0],edge[1],'solid'))
-
-		dot_str += ("}\n")
-		
-		return dot_str
-
-	def save(self, filename):
-
-		dot_str = self.generate_dot()
-		s = Source(dot_str, filename=filename, format="png")
-
-	def show(self):
-
-		dot_str = self.generate_dot()
-		s = Source(dot_str, filename='test.gv', format="png")
-		s.view()
-
-	def convert2LTL(self):
-		'''
-			returns the LTL equivalent of the DT
-		'''
-		if self.label == formula_true or self.label== formula_false:
-			return self.label
-
-		if self.left.label == Formula('true'):
-			first_disjunct = self.label
-		elif self.left.label == Formula('false'):
-			first_disjunct = None
-		else:
-			first_disjunct = Formula(['&', self.label, self.left.convert2LTL()])
-
-		if self.right.label == Formula('true'):
-			second_disjunct = self.label
-		elif self.right.label == Formula('false'):
-			second_disjunct = None
-		else:
-			second_disjunct = Formula(['&', Formula(['!', self.label]), self.right.convert2LTL()])
-
-
-		if first_disjunct != None and second_disjunct != None:
-			final_formula = Formula(['|', first_disjunct, second_disjunct])	
-		elif first_disjunct == None:
-			final_formula = second_disjunct
-		elif second_disjunct == None:
-			final_formula = first_disjunct
-
-		return final_formula
-
-	def __eq__(self, other):
-		if other == None:
-			return False
-		else:
[…]
