+++ Scarlet_ltl.egg-info/SOURCES.txt
@@ -4,14 +4,12 @@ pyproject.toml
 Scarlet/__init__.py
 Scarlet/booleanSubsetCover.py
 Scarlet/convert2dfa.py
-Scarlet/decisionTree.py
 Scarlet/directed_ltl.py
 Scarlet/formulaTree.py
 Scarlet/genBenchmarks.py
 Scarlet/inferLTL.py
 Scarlet/ltllearner.py
 Scarlet/sample.py
-Scarlet/wrapper.py
 Scarlet_ltl.egg-info/PKG-INFO
 Scarlet_ltl.egg-info/SOURCES.txt
 Scarlet_ltl.egg-info/dependency_links.txt
