### 2023-10-02 rajarshi008/Scarlet#8 @ritamraha
The errors were caused due to inconsistency in the Python package of Scarlet in Pypi and the sources in Git Hub. We have fixed the issues now with the Python package. To test, the Scarlet Python package has to be upgraded or reinstalled to the latest version. Alternatively, Scarlet can be installed via the command line and then also be tested following the instructions in README.
