### 2023-12-09 #5052 @Smattr
Thanks for the bump. Unbelievable as it may sound, I had not forgotten about this and it was still in my inbox. I just have not been able to find time to get to it. I’ll take another look now.

> > The mention of artificial intelligence seems a little odd when nothing here seems specific to or especially suited to AI. I guess AI is a hot topic right now, so is this some keyword gaming? (to which I don’t object, but just trying to understand if I’ve missed a connection here.)
>
> There is a massive surge of interest in formally representing knowledge of intelligent agents in the AI community. This is particularly prominent in the sub-domains of AI known as Explainable AI (XAI) and Knowledge Representation (KR), where an important problem is to identify formalisms that are both human-interpretable and unambiguous for representing properties and behavior of intelligent agents. As seen in several papers in recent AI research [1][2][3][4][5][6], LTL and other temporal logics have been proposed as rigorous formalisms for formally representing knowledge in applications such as planning, reinforcement learning, synthesis, etc.

I read this as a “yes” to my question. Thanks for the clarification.

> > Given the generalised audience of JOSS, this could probably provide more context on the domain (formal methods, proofs, etc). To formal methods practitioners, the utility is going to be clear. But to everyone else, I imagine they’re wondering what you would use things like LTL and SAT for.
>
> While Scarlet can be used in Formal Methods applications such as specification mining, it was designed mainly to cater to AI applications, as motivated in the [Summary](https://github.com/rajarshi008/Scarlet/blob/main/paper.md#summary). Moreover, the variant of LTL used in Scarlet, LTL on finite traces [7], is better suited for AI applications where the trajectories are typically of finite length.

I don’t understand how this addresses my feedback. I was not requesting an inline summary of AI or LTL. I, the reviewer, understand these topics. I was requesting more background information be inserted into the paper. I’ll take a re-read and see if anything has been added.
