### 2023-12-10 #5052 @Smattr
Current state looks good. I’ve ticked off the last few items.

IMHO it is not possible to understand this tool or its context from the submitted paper alone. Most substantial aspects of both (tool and context) are just references to other (pay-walled) papers. Nevertheless the addition of this to JOSS seems worthwhile.
