### 2023-12-11 #5052 @ritamraha
> @Smattr @JDRomano2 thank you for completing your respective reviews. I appreciate your thoroughness!
>
> @ritamraha Looks like everything is set. Congratulations! I need one final formality to be done from your side. Please make a tagged release of the project and archive it. Then report the version number and archive DOI here. Once I have that, I will send this for publication. Cheers!

We thank both reviewers for providing their constructive feedback. @adi3, thank you very much for your efforts.

We have made a tagged release of the project  named ["Joss Release"](https://github.com/rajarshi008/Scarlet/releases/tag/v1.0.0) with version 1.0.0. The DOI of the archived release is (https://doi.org/10.5281/zenodo.10357545).
