+++ paper.bib
@@ -29,7 +29,7 @@
   title     = {{ARSENAL:} Automatic Requirements Specification Extraction from Natural Language},
   booktitle = {{NASA} Formal Methods, NFM},
   year      = {2016},
-  doi       = {10.1007/978-3-319-40648-0\_4},
+  doi       = {10.1007/978-3-319-40648-0_4},
 }
 
 @inproceedings{FijalkowLagarde21,
@@ -79,7 +79,7 @@
   booktitle = {Formal Methods in Computer Aided Design, FMCAD},
   pages     = {104--112},
   year      = {2020},
-  doi       = {10.34727/2020/isbn.978-3-85448-042-6\_17},
+  doi       = {10.34727/2020/isbn.978-3-85448-042-6_17},
 }
 
   
@@ -96,7 +96,8 @@
   title     = {Learning Interpretable Models Expressed in Linear Temporal Logic}, 
   journal   = {International Conference on Automated Planning and Scheduling, ICAPS}, 
   year      = {2019}, 
-  url       = {https://ojs.aaai.org/index.php/ICAPS/article/view/3529}, 
+  url       = {https://ojs.aaai.org/index.php/ICAPS/article/view/3529},
+  doi       = {10.1609/icaps.v29i1.3529}
 }
 
 @inproceedings{LPB15,
@@ -137,7 +138,7 @@
   title     = {cvc4sy: Smart and Fast Term Enumeration for Syntax-Guided Synthesis},
   booktitle = {Computer-Aided Verification, CAV},
   year      = {2019},
-  doi       = {10.1007/978-3-030-25543-5\_5},
+  doi       = {10.1007/978-3-030-25543-5_5},
 }
 
 @inproceedings{RahaRFN22,
@@ -154,4 +155,33 @@
   publisher = {Springer},
   year      = {2022},
   doi       = {10.1007/978-3-030-99524-9_14},
+}
+
+@article{abs-2310-13778,
+  author       = {Rajarshi Roy and
+                  Daniel Neider},
+  title        = {Inferring Properties in Computation Tree Logic},
+  journal      = {CoRR},
+  volume       = {abs/2310.13778},
+  year         = {2023},
+  url          = {https://doi.org/10.48550/arXiv.2310.13778},
+  doi          = {10.48550/ARXIV.2310.13778},
+  eprinttype    = {arXiv},
+  eprint       = {2310.13778},
+  timestamp    = {Mon, 30 Oct 2023 08:58:20 +0100}
+}
+
+@article{abs-2310-17410,
+  author       = {Ritam Raha and
+                  Rajarshi Roy and
+                  Nathana{\"{e}}l Fijalkow and
+                  Daniel Neider and
+                  Guillermo A. P{\'{e}}rez},
+  title        = {Synthesizing Efficiently Monitorable Formulas in Metric Temporal Logic},
+  journal      = {CoRR},
+  volume       = {abs/2310.17410},
+  year         = {2023},
+  url          = {https://doi.org/10.48550/arXiv.2310.17410},
+  doi          = {10.48550/ARXIV.2310.17410},
+  eprinttype    = {arXiv}
 }
