+++ paper.bib
@@ -126,6 +126,7 @@
   title     = {{SYSLITE:} Syntax-Guided Synthesis of {PLTL} Formulas from Finite
                Traces},
   booktitle = {Formal Methods in Computer Aided Design, {FMCAD}},
+  doi          = {10.34727/2020/ISBN.978-3-85448-042-6_16},
   year      = {2020},
 }
 
