### 2023-12-21 #5052 @rajarshi008
Hey @adi3, we have fixed the issues with the references and then rearchived. We updated the release  ["Joss Release"](https://github.com/rajarshi008/Scarlet/releases/tag/v1.0.0) with version 1.0.0 and the doi (https://doi.org/10.5281/zenodo.10419514).
